+++ clinica/pipelines/engine_pydra.py
@@ -1,13 +1,16 @@
+from pathlib import PurePath
 import nipype.interfaces.io as nio
 import pydra
 from pydra import Submitter, Workflow
 
 import clinica.pipelines.engine_pydra_utils as pu
-from clinica.pipelines.interfaces import bids_reader_task, bids_writer_task
+from clinica.pipelines.interfaces import bids_writer
 
 
 class Pipeline:
-    def __init__(self, name, pipeline_in_dir, pipeline_out_dir):
+    def __init__(
+        self, name: str, pipeline_in_dir: PurePath, pipeline_out_dir: PurePath
+    ):
         """
         :pipeline_in_dir: a BIDS compliant input directory
         """
@@ -15,7 +18,7 @@ class Pipeline:
         self.pipeline_out_dir = pipeline_out_dir
         self.name = name
 
-    def build_workflow(self, core_workflow):
+    def build_workflow(self, core_workflow: Workflow):
         """
         Construct Clinica compliant pipeline
         :core_workflow: a pydra workflow with the core interfaces
@@ -39,22 +42,9 @@ class Pipeline:
             bids_data = getattr(self.input_workflow.lzout, field)
             setattr(core_workflow.inputs, field, bids_data)
 
-        #
         # @TODO: define condition on split if multiple fields
-        #
 
         self.workflow.add(core_workflow.split(field))
-
-        @pydra.mark.task
-        @pydra.mark.annotate({"return": {"output_file": str}})
-        def bids_writer(output_file):
-            """
-            (Dummy) Task to echo files generated by the core workflow
-            """
-
-            print("printing core vars in output_workflow", output_file)
-
-            return "dummy_output_string"
 
         for i, field in enumerate(pu.list_out_fields(core_workflow)):
 
@@ -73,11 +63,11 @@ class Pipeline:
         return self.workflow
 
     @property
-    def input_workflow(self):
+    def input_workflow(self) -> Workflow:
         return self._input_workflow
 
     @input_workflow.setter
-    def input_workflow(self, query_bids, name: str = "input"):
+    def input_workflow(self, query_bids: dict, name: str = "input"):
 
         """Generic Input workflow
         :query_bids: input to BIDSDataGrabber
@@ -114,5 +104,5 @@ class Pipeline:
             submitter(self.workflow)
 
         results = self.workflow.result(return_inputs=True)
-
+        # @TODO: decide where to store results
         print(results)
