+++ clinica/pipelines/interfaces.py
@@ -3,27 +3,13 @@ import pydra
 from pydra.tasks.nipype1.utils import Nipype1Task
 
 
-def bids_reader_task(input_dir, query_bids):
-    """
-    Return pydra Nipype1Task for querying bids dataset
-    """
-
-    bids_data_grabber = nio.BIDSDataGrabber(
-        output_query=query_bids,
-    )
-
-    task_bids_data_grabber = Nipype1Task(
-        name="pydra_bids_data_reader", base_dir=input_dir, interface=bids_data_grabber
-    )
-
-    return task_bids_data_grabber
-
-
 @pydra.mark.task
-def bids_writer_task(input_file, output_dir):
+@pydra.mark.annotate({"return": {"output_file": str}})
+def bids_writer(output_file):
     """
-    Task to write files to output_dir
+    (Dummy) Task to echo files generated by the core workflow
     """
-    import subprocess
 
-    subprocess.run(["cp", input_file, output_dir])
+    print("printing core vars in output_workflow", output_file)
+
+    return "dummy_output_string"
